Add VTY option to add forbidden PLMN network manually
The given network is appended and also stored in the SIM card.

This is useful to block networks that would detach SIM cards with unknown
(test) PLMN. Then the mobile will not register to that network, as long it
is kept in the blocked list.

Note that only the last four forbidden PLMNs will be stored in the SIM card.